Implement actual clock frequency with the de0 nano clock
Add the actual clock frequency which will be mapped using the quartus prime. I am planning to use the PLL of the board to measure the frequency of every component this will be used in order to work with both uart and i2c at the same time. Added std_types c like signals but till i fix the unsigned warnings I get I will not use this